mcp-travel-agent/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-travel-agent-backend"
version = "0.1.0"
description = "AI-ready backend API for mcp-travel-agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.3.0",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "google-generativeai>=0.8.0",
  "langgraph>=0.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "mypy>=1.10",
  "ruff>=0.5"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_configs = true
